dev-requirements.txt
-r requirements.txt

# Stuff only needed in dev:
nose
coverage

Flask-Testing==0.4.2
blinker==1.3
